criando classe de metodos e alteracao do nome das classes
